Add files via upload
This is an example script that can be used to collect the outcome variables of interest from the CSV files generated from the NeoTapp application. The code must be tailored to the particular data collection and the outcome variables of interest to a specific project, but the example showcase how this can be done in a relatively easy way.